Rename module `id-syntax-default` -> `doc-id-default`.
