+++ local-modules/@bayou/config-common-default/IdSyntax.js
@@ -2,12 +2,12 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { DefaultIdSyntax } from '@bayou/id-syntax-default';
+import { DefaultIdSyntax } from '@bayou/doc-id-default';
 import { BaseIdSyntax } from '@bayou/config-common';
 
 /**
  * Utility functionality regarding ID strings. This implementation uses the
- * default definitions provided by {@link @bayou/id-syntax-default}.
+ * default definitions provided by {@link @bayou/doc-id-default}.
  */
 export default class IdSyntax extends BaseIdSyntax {
   /**
