+++ local-modules/@bayou/doc-id-default/README.md
@@ -1,4 +1,4 @@
-@bayou/id-syntax-default
+@bayou/doc-id-default
 ========================
 
 Default definitions of syntax for all the kinds of IDs used by the system. This
