+++ local-modules/@bayou/file-store-local/LocalFileStore.js
@@ -9,7 +9,7 @@ import { Codec } from '@bayou/codec';
 import { Dirs } from '@bayou/env-server';
 import { BaseFileStore, FileCache } from '@bayou/file-store';
 import { TheModule as fileStoreOt_TheModule } from '@bayou/file-store-ot';
-import { DefaultIdSyntax } from '@bayou/id-syntax-default';
+import { DefaultIdSyntax } from '@bayou/doc-id-default';
 import { Logger } from '@bayou/see-all';
 
 import LocalFile from './LocalFile';
@@ -86,7 +86,7 @@ export default class LocalFileStore extends BaseFileStore {
 
   /**
    * Implementation as required by the superclass. This implementation defers
-   * to the module {@link @bayou/id-syntax-default}.
+   * to the module {@link @bayou/doc-id-default}.
    *
    * @param {string} fileId The alleged file ID.
    * @returns {boolean} `true` if `fileId` is a syntactically valid file ID, or
